powerpc/32: Split head_32.h into head_40x.h and head_6xx_8xx.h
book3s/32 (aka 6xx) and 8xx head will be reworked to re-enable MMU
earlier.

Split 40x head.h out so that we can keep 40x as is until it
is phased out.

There is no plan to implement VMAP stack on 40x on the near future
so remove everything related.

Signed-off-by: Christophe Leroy <christophe.leroy@csgroup.eu>